pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "suno-cli"
version = "0.1.0"
description = "CLI tool for generating songs with Suno AI via sunoapi.org"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
authors = [
    { name = "Simon Lauger", email = "simon@lauger.name" }
]
keywords = ["suno", "ai", "music", "generation", "cli", "song"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Sound/Audio",
]

dependencies = [
    "requests>=2.31.0",
    "click>=8.1.0",
    "rich>=13.0.0",
    "mutagen>=1.47.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]
youtube = [
    "google-auth-oauthlib>=1.0.0",
    "google-api-python-client>=2.100.0",
]

[project.scripts]
suno = "suno_cli.cli:cli"

[project.urls]
Homepage = "https://github.com/slauger/suno-cli"
Repository = "https://github.com/slauger/suno-cli"
Issues = "https://github.com/slauger/suno-cli/issues"

[tool.hatch.build.targets.wheel]
packages = ["src/suno_cli"]

[tool.black]
line-length = 100
target-version = ['py38']

[tool.ruff]
line-length = 100
target-version = "py38"
